reconnect delay, remove writer goroutine, refactor
